+++ packages/orq-rc/src/orq_ai_sdk/models/createidentityop.py
@@ -137,7 +137,7 @@ class CreateIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-23T13:57:49.986Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-23T14:47:21.209Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")