+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ5CNZX9WA5BE36T59J5AD7R"
+        "01KJ5FGNHAZASNSA2XE3BDYNA5"
     )
     r"""The unique identifier of the data source"""