+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ5CNZXD3QRSTE01M8QGY17X"
+        "01KJ5FGNHF29AJFMBAPYVFXJNX"
     )
     r"""The unique identifier of the data source"""